# docs/ai-tools-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DashAdmin AI Tools API",
    "version": "2.0.0",
    "description": "API for AI agents (n8n) to query DashAdmin data. Requires bot authentication via headers."
  },
  "servers": [
    {
      "url": "https://www.mydashadmin.ru",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "BotAuth": []
    }
  ],
...
  "paths": {
    "/api/ai-tools/select-club": {
      "get": {
        "summary": "Get Available Clubs",
        "description": "Returns the list of clubs available to the authenticated user and their current selection.",
        "operationId": "getAvailableClubs",
        "responses": {
          "200": {
            "description": "Club information",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ClubSelection" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Select Club",
        "description": "Sets the active club for subsequent AI queries.",
        "operationId": "selectClub",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SelectClubRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Club selected successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ClubSelection" }
              }
            }
          }
        }
      }
    },
    "/api/ai-tools/revenue": {
      "get": {
        "summary": "Get Revenue Report",
        "description": "Fetches revenue data for the selected club. Supports various time periods.",
        "operationId": "getRevenue",
        "parameters": [
          {
            "name": "clubId",
            "in": "query",
            "required": false,
            "description": "Override club ID (defaults to selected club)",
            "schema": { "type": "string" }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "description": "Predefined time period",
            "schema": {
              "type": "string",
              "enum": ["today", "yesterday", "last7days", "last30days", "this_month"]
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "description": "Custom start date (YYYY-MM-DD)",
            "schema": { "type": "string" }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "description": "Custom end date (YYYY-MM-DD)",
            "schema": { "type": "string" }
          },
          {
            "name": "shiftType",
            "in": "query",
            "required": false,
            "description": "Filter by shift type",
            "schema": { "type": "string", "enum": ["DAY", "NIGHT"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RevenueReport" }
              }
            }
          }
        }
      }
    },
    "/api/ai-tools/shifts-summary": {
      "get": {
        "summary": "Get Shifts Summary",
        "description": "Fetches summary statistics for shifts including hours, counts, and employee data.",
        "operationId": "getShiftsSummary",
        "parameters": [
          {
            "name": "clubId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["today", "yesterday", "last7days", "last30days", "this_month"]
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filter by shift status",
            "schema": { "type": "string", "enum": ["ACTIVE", "CLOSED"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Shifts summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShiftsSummary" }
              }
            }
          }
        }
      }
    },
    "/api/ai-tools/employees-list": {
      "get": {
        "summary": "Get Employees List",
        "description": "Returns a list of employees for the club with their basic information.",
        "operationId": "getEmployeesList",
        "parameters": [
          {
            "name": "clubId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "includeDismissed",
            "in": "query",
            "required": false,
            "description": "Include dismissed employees",
            "schema": { "type": "string", "enum": ["true", "false"] }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of results",
            "schema": { "type": "integer", "default": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Employees list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmployeesList" }
              }
            }
          }
        }
      }
    },
    "/api/ai-tools/employee-hours": {
      "get": {
        "summary": "Get Employee Hours",
        "description": "Returns hours worked by employees. Can search by employee ID, name, or return all employees.",
        "operationId": "getEmployeeHours",
        "parameters": [
          {
            "name": "clubId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "employeeId",
            "in": "query",
            "required": false,
            "description": "UUID of specific employee",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "employeeName",
            "in": "query",
            "required": false,
            "description": "Search employees by name (partial match)",
            "schema": { "type": "string" }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["today", "yesterday", "last7days", "last30days", "this_month"]
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee hours data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmployeeHours" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BotAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Messenger-User-Id",
        "description": "Bot authentication via linked user account"
      }
    },
    "schemas": {
      "ClubSelection": {
        "type": "object",
        "properties": {
          "selected_club": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": { "type": "integer" },
              "name": { "type": "string" }
            }
          },
          "available_clubs": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "name": { "type": "string" }
              }
            }
          }
        }
      },
      "SelectClubRequest": {
        "type": "object",
        "required": ["clubId"],
        "properties": {
          "clubId": { "type": "string", "description": "ID of club to select" }
        }
      },
      "RevenueReport": {
        "type": "object",
        "properties": {
          "club_id": { "type": "integer" },
          "club_name": { "type": "string" },
          "period": {
            "type": "object",
            "properties": {
              "start": { "type": "string" },
              "end": { "type": "string" },
              "label": { "type": "string" }
            }
          },
          "revenue": {
            "type": "object",
            "properties": {
              "total": { "type": "number" },
              "expenses": { "type": "number" },
              "net": { "type": "number" },
              "cash": { "type": "number" },
              "card": { "type": "number" }
            }
          },
          "shifts": {
            "type": "object",
            "properties": {
              "count": { "type": "integer" },
              "total_hours": { "type": "number" },
              "employees": { "type": "integer" }
            }
          }
        }
      },
      "ShiftsSummary": {
        "type": "object",
        "properties": {
          "club_id": { "type": "integer" },
          "club_name": { "type": "string" },
          "period": { "$ref": "#/components/schemas/RevenueReport/properties/period" },
          "summary": {
            "type": "object",
            "properties": {
              "total_shifts": { "type": "integer" },
              "total_hours": { "type": "number" },
              "average_duration_hours": { "type": "number" },
              "day_shifts": { "type": "integer" },
              "night_shifts": { "type": "integer" },
              "active_shifts": { "type": "integer" },
              "closed_shifts": { "type": "integer" },
              "unique_employees": { "type": "integer" }
            }
          }
        }
      },
      "EmployeesList": {
        "type": "object",
        "properties": {
          "club_id": { "type": "integer" },
          "club_name": { "type": "string" },
          "total_count": { "type": "integer" },
          "employees": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "name": { "type": "string" },
                "phone": { "type": "string" },
                "role": { "type": "string" },
                "hired_at": { "type": "string" },
                "is_active": { "type": "boolean" },
                "is_dismissed": { "type": "boolean" },
                "show_in_schedule": { "type": "boolean" }
              }
            }
          }
        }
      },
      "EmployeeHours": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "club_id": { "type": "integer" },
              "club_name": { "type": "string" },
              "period": { "$ref": "#/components/schemas/RevenueReport/properties/period" },
              "employee": {
                "type": "object",
                "properties": {
                  "id": { "type": "string" },
                  "name": { "type": "string" }
                }
              },
              "stats": {
                "type": "object",
                "properties": {
                  "shifts_count": { "type": "integer" },
                  "total_hours": { "type": "number" },
                  "avg_shift_hours": { "type": "number" },
                  "total_revenue": { "type": "number" }
                }
              }
            }
          },
          {
            "type": "object",
            "properties": {
              "club_id": { "type": "integer" },
              "club_name": { "type": "string" },
              "period": { "$ref": "#/components/schemas/RevenueReport/properties/period" },
              "search_query": { "type": "string" },
              "employees": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "name": { "type": "string" },
                    "shifts_count": { "type": "integer" },
                    "total_hours": { "type": "number" },
                    "avg_shift_hours": { "type": "number" },
                    "total_revenue": { "type": "number" }
                  }
                }
              }
            }
          }
        ]
      }
    }
  }
}